Update figma plugin API calls to Async (sync were deprecated)
